incremental changes

+++ include/nlohmann/detail/input/numerizer.hpp
@@ -31,13 +31,13 @@ struct numerizer
     }
 
     JSON_HEDLEY_NON_NULL(2)
-    static unsigned long long int strtoull(const char* str, char** endptr, int base)
+    static uint64_t strtoull(const char* str, char** endptr, int base)
     {
         return std::strtoull(str, endptr, base);
     }
 
     JSON_HEDLEY_NON_NULL(2)
-    static long long int strtoll(const char* str, char** endptr, int base)
+    static int64_t strtoll(const char* str, char** endptr, int base)
     {
         return std::strtoll(str, endptr, base);
     }
